first_stage.py - convert sw, sh to ints

+++ backend-python/detector/first_stage.py
@@ -25,7 +25,7 @@ def run_first_stage(image, net, scale, threshold):
 
     # scale the image and convert it to a float array
     width, height = image.size
-    sw, sh = math.ceil(width*scale), math.ceil(height*scale)
+    sw, sh = int(math.ceil(width*scale)), int(math.ceil(height*scale))
     img = image.resize((sw, sh), Image.BILINEAR)
     img = np.asarray(img, 'float32')
